Move populate_module_imports out of ruff_linter
